Add files via upload
This python script expects a directory loaded with JPEG images of foundation patches extracted from the web page of a cosmetic manufacturer.  The directory is chosen by selecting one of the files in it using the usual sort of file browser.  The script will put in the directory a file called rgb.txt.  The first line is the directory path.  The remaining lines list the cosmetic name and RGB reflectances for each of the patches.